vectorization-server/requirements.txt
fastapi
uvicorn
kafka-python
requests
torch
torchvision
Pillow
python-dotenv
sentence_transformers